Refactor Chat Window
