+++ src/components/Buttons/index.ts
@@ -1,5 +1,6 @@
-import IconButton from './IconButton';
-import TextButton from './TextButton';
 import Button from './Button';
+import IconButton from './IconButton';
+import StatusButton from './StatusButton';
+import TextButton from './TextButton';
 
-export { IconButton, TextButton, Button };
+export { Button, IconButton, StatusButton, TextButton };
